Add barrel files for SparkSwap Website components
